Release 1.2.0

+++ scripts/Modules/SPSCleanDependencies.util.psm1
@@ -1,13 +1,7 @@
-# Ensure the script is running with administrator privileges
-if (-NOT ([Security.Principal.WindowsPrincipal] [Security.Principal.WindowsIdentity]::GetCurrent()).IsInRole([Security.Principal.WindowsBuiltInRole] 'Administrator')) {
-    Throw "Administrator rights are required. Please re-run this script as an Administrator."
-}
-# Setting power management plan to High Performance"
-Start-Process -FilePath "$env:SystemRoot\system32\powercfg.exe" -ArgumentList '/s 8c5e7fda-e8bf-4a96-9a85-a6e23a8c635c' -NoNewWindow
 function Get-SPSInstalledProductVersion {
     [OutputType([System.Version])]
     param ()
-  
+
     $pathToSearch = 'C:\Program Files\Common Files\microsoft shared\Web Server Extensions\*\ISAPI\Microsoft.SharePoint.dll'
     $fullPath = Get-Item $pathToSearch -ErrorAction SilentlyContinue | Sort-Object { $_.Directory } -Descending | Select-Object -First 1
     if ($null -eq $fullPath) {
@@ -17,25 +11,38 @@ function Get-SPSInstalledProductVersion {
         return (Get-Command $fullPath).FileVersionInfo
     }
 }
-# Load SharePoint Powershell Snapin or Import-Module
-try {
-    $installedVersion = Get-SPSInstalledProductVersion
-    if ($installedVersion.ProductMajorPart -eq 15 -or $installedVersion.ProductBuildPart -le 12999) {
-        if ($null -eq (Get-PSSnapin -Name Microsoft.SharePoint.PowerShell -ErrorAction SilentlyContinue)) {
-            Add-PSSnapin Microsoft.SharePoint.PowerShell
+
+# Import-time prelude (admin check, power plan, SharePoint snap-in load).
+# Skipped when $env:SPSCD_SKIP_PRELUDE is set, so the module can be imported on
+# CI / non-SharePoint hosts (e.g. Pester) without elevation or SharePoint installed.
+if (-not $env:SPSCD_SKIP_PRELUDE) {
+    # Ensure the script is running with administrator privileges
+    if (-NOT ([Security.Principal.WindowsPrincipal] [Security.Principal.WindowsIdentity]::GetCurrent()).IsInRole([Security.Principal.WindowsBuiltInRole] 'Administrator')) {
+        Throw "Administrator rights are required. Please re-run this script as an Administrator."
+    }
+    # Setting power management plan to High Performance"
+    Start-Process -FilePath "$env:SystemRoot\system32\powercfg.exe" -ArgumentList '/s 8c5e7fda-e8bf-4a96-9a85-a6e23a8c635c' -NoNewWindow
+
+    # Load SharePoint Powershell Snapin or Import-Module
+    try {
+        $installedVersion = Get-SPSInstalledProductVersion
+        if ($installedVersion.ProductMajorPart -eq 15 -or $installedVersion.ProductBuildPart -le 12999) {
+            if ($null -eq (Get-PSSnapin -Name Microsoft.SharePoint.PowerShell -ErrorAction SilentlyContinue)) {
+                Add-PSSnapin Microsoft.SharePoint.PowerShell
+            }
+        }
+        else {
+            Import-Module SharePointServer -Verbose:$false -WarningAction SilentlyContinue -DisableNameChecking
         }
     }
-    else {
-        Import-Module SharePointServer -Verbose:$false -WarningAction SilentlyContinue -DisableNameChecking
-    }
-}
-catch {
-    # Handle errors during retrieval of Installed Product Version
-    $catchMessage = @"
+    catch {
+        # Handle errors during retrieval of Installed Product Version
+        $catchMessage = @"
 Failed to get installed Product Version for $($env:COMPUTERNAME)
 Exception: $($_.Exception.Message)
 "@
-    Write-Error -Message $catchMessage
+        Write-Error -Message $catchMessage
+    }
 }
 
 # Initialize jSON Object
@@ -69,6 +76,13 @@ class SPMissingWebPartInfo {
     [System.String]$WebPartID
     [System.String]$Message
     [System.String]$Remedy
+    [System.String]$ClassName
+    [System.String]$StorageKey
+    [System.String]$SiteID
+    [System.String]$WebID
+    [System.String]$ListID
+    [System.String]$DirName
+    [System.String]$LeafName
 }
 class SPMissingSetupFileInfo {
     [System.String]$Database
@@ -217,6 +231,64 @@ FROM EventReceivers (NOLOCK) where Assembly = '$($AssemblyInfo)'
     return $tbSQLmissingAssemblies
 }
 
+function Get-SQLMissingWebPartInfo {
+    param
+    (
+        [Parameter()]
+        [System.String]
+        $DatabaseName,
+
+        [Parameter()]
+        [System.String]
+        $DatabaseServer,
+
+        [Parameter()]
+        [System.String]
+        $ClassID
+    )
+
+    class SQLMissingWebPartInfo {
+        [System.String]$StorageKey
+        [System.String]$ClassName
+        [System.String]$SiteID
+        [System.String]$WebID
+        [System.String]$ListID
+        [System.String]$DirName
+        [System.String]$LeafName
+    }
+    $tbSQLmissingWebParts = New-Object -TypeName System.Collections.ArrayList
+    try {
+        $sqlQuery =
+        @"
+USE $($DatabaseName)
+SELECT wp.tp_ID, d.SiteId, d.WebId, d.ListId, d.DirName, d.LeafName, wp.tp_Class
+FROM AllDocs d WITH (NOLOCK)
+INNER JOIN AllWebParts wp WITH (NOLOCK) ON wp.tp_PageUrlID = d.Id
+INNER JOIN AllWebs w WITH (NOLOCK) ON d.WebId = w.id
+WHERE wp.tp_WebPartTypeId = '$($ClassID)'
+"@
+
+        $invokeSQLQueries = Invoke-Sqlcmd -Query $sqlQuery `
+            -ServerInstance "$($DatabaseServer)"
+
+        foreach ($invokeSQLQuery in $invokeSQLQueries) {
+            [void]$tbSQLmissingWebParts.Add([SQLMissingWebPartInfo]@{
+                    StorageKey = $invokeSQLQuery.tp_ID;
+                    ClassName  = $invokeSQLQuery.tp_Class;
+                    SiteID     = $invokeSQLQuery.SiteId;
+                    WebID      = $invokeSQLQuery.WebId;
+                    ListID     = $invokeSQLQuery.ListId;
+                    DirName    = $invokeSQLQuery.DirName;
+                    LeafName   = $invokeSQLQuery.LeafName;
+                })
+        }
+    }
+    catch {
+        return $_
+    }
+    return $tbSQLmissingWebParts
+}
+
 function Get-SQLMissingConfiguration {
     param
     (
@@ -298,13 +370,37 @@ function Get-SPSMissingServerDependencies {
                     if ($null -ne $missingWebParts) {
                         foreach ($missingWebPart in $missingWebParts) {
                             $webPartID = ([regex]::Matches($missingWebPart.Message, '[0-9a-fA-F]{8}\-[0-9a-fA-F]{4}\-[0-9a-fA-F]{4}\-[0-9a-fA-F]{4}\-[0-9a-fA-F]{12}')).value
-                            [void]$tbSPmissingWebParts.Add([SPMissingWebPartInfo]@{
-                                    Database  = "$($spContentDB.Name)";
-                                    Category  = $missingWebPart.Category;
-                                    WebPartID = $webPartID
-                                    Message   = $missingWebPart.Message;
-                                    Remedy    = $missingWebPart.Remedy;
-                                })
+                            $sqlMissingWebParts = Get-SQLMissingWebPartInfo -DatabaseName "$($spContentDB.Name)" `
+                                -DatabaseServer "$($spContentDB.Server)" `
+                                -ClassID "$($webPartID)"
+
+                            if ($null -ne $sqlMissingWebParts -and $sqlMissingWebParts.Count -gt 0) {
+                                foreach ($sqlMissingWebPart in $sqlMissingWebParts) {
+                                    [void]$tbSPmissingWebParts.Add([SPMissingWebPartInfo]@{
+                                            Database   = "$($spContentDB.Name)";
+                                            Category   = $missingWebPart.Category;
+                                            WebPartID  = $webPartID;
+                                            Message    = $missingWebPart.Message;
+                                            Remedy     = $missingWebPart.Remedy;
+                                            ClassName  = $sqlMissingWebPart.ClassName;
+                                            StorageKey = $sqlMissingWebPart.StorageKey;
+                                            SiteID     = $sqlMissingWebPart.SiteID;
+                                            WebID      = $sqlMissingWebPart.WebID;
+                                            ListID     = $sqlMissingWebPart.ListID;
+                                            DirName    = $sqlMissingWebPart.DirName;
+                                            LeafName   = $sqlMissingWebPart.LeafName;
+                                        })
+                                }
+                            }
+                            else {
+                                [void]$tbSPmissingWebParts.Add([SPMissingWebPartInfo]@{
+                                        Database  = "$($spContentDB.Name)";
+                                        Category  = $missingWebPart.Category;
+                                        WebPartID = $webPartID;
+                                        Message   = $missingWebPart.Message;
+                                        Remedy    = $missingWebPart.Remedy;
+                                    })
+                            }
                         }
                     }
                     if ($null -ne $missingSetupFiles) {
@@ -445,6 +541,7 @@ function Get-SPSMissingServerDependencies {
 }
 
 function Remove-SPSMissingFeature {
+    [CmdletBinding(SupportsShouldProcess = $true)]
     param
     (
         [Parameter()]
@@ -475,7 +572,9 @@ function Remove-SPSMissingFeature {
             ForEach ($web in $Site.AllWebs) {
                 if ($web.Features[$featureID]) {
                     Write-Output "`nFound Feature $featureID in web:"$Web.Url"`nRemoving feature"
-                    $web.Features.Remove($featureID, $true)
+                    if ($PSCmdlet.ShouldProcess($Web.Url, "Remove feature $featureID")) {
+                        $web.Features.Remove($featureID, $true)
+                    }
                 }
                 else {
                     Write-Output "`nDid not find feature $featureID in web:" $Web.Url
@@ -484,7 +583,9 @@ function Remove-SPSMissingFeature {
             #Remove the feature from the site collection
             if ($Site.Features[$featureID]) {
                 Write-Output "`nFound feature $featureID in site:"$site.Url"`nRemoving Feature"
-                $site.Features.Remove($featureID, $true)
+                if ($PSCmdlet.ShouldProcess($site.Url, "Remove feature $featureID")) {
+                    $site.Features.Remove($featureID, $true)
+                }
             }
             else {
                 Write-Output "Did not find feature $featureID in site:" $site.Url
@@ -500,6 +601,7 @@ function Remove-SPSMissingFeature {
 }
 
 function Remove-SPSMissingSetupFile {
+    [CmdletBinding(SupportsShouldProcess = $true)]
     param
     (
         [Parameter()]
@@ -537,9 +639,11 @@ function Remove-SPSMissingSetupFile {
                 $file = $web.GetFile([GUID]$FileID)
                 if ($null -ne $file.ServerRelativeUrl) {
                     $filelocation = "{0}{1}" -f ($site.WebApplication.Url).TrimEnd("/"), $file.ServerRelativeUrl
-                    Write-Host "Found file location:" $filelocation
+                    Write-Output "Found file location: $filelocation"
                     #Delete the file, the Delete() method bypasses the recycle bin
-                    $file.Delete()
+                    if ($PSCmdlet.ShouldProcess($filelocation, "Delete setup file $FileID")) {
+                        $file.Delete()
+                    }
                     $web.dispose()
                     $site.dispose()
                 }
@@ -560,7 +664,106 @@ function Remove-SPSMissingSetupFile {
     }
 }
 
+function Remove-SPSMissingWebPart {
+    [CmdletBinding(SupportsShouldProcess = $true)]
+    param
+    (
+        [Parameter()]
+        [System.String]
+        $Database,
+
+        [Parameter()]
+        [System.String]
+        $WebPartID,
+
+        [Parameter()]
+        [System.String]
+        $StorageKey,
+
+        [Parameter()]
+        [System.String]
+        $SiteID,
+
+        [Parameter()]
+        [System.String]
+        $WebID,
+
+        [Parameter()]
+        [System.String]
+        $DirName,
+
+        [Parameter()]
+        [System.String]
+        $LeafName
+    )
+
+    try {
+        Write-Output '-----------------------------------------------'
+        Write-Output 'Removing Missing WebPart Dependencies of:'
+        Write-Output " * Database: $Database"
+        Write-Output " * WebPartID (class): $WebPartID"
+        Write-Output " * StorageKey: $StorageKey"
+        Write-Output " * SiteID: $SiteID"
+        Write-Output " * WebID: $WebID"
+        Write-Output " * Page: $DirName/$LeafName"
+        Write-Output '-----------------------------------------------'
+
+        if ([string]::IsNullOrEmpty($StorageKey)) {
+            Write-Output 'StorageKey is empty - no specific WebPart instance to remove. Re-run the audit to collect per-page locations.'
+            return
+        }
+
+        $site = Get-SPSite -Limit All -Identity $SiteID -ErrorAction SilentlyContinue
+        if ($null -ne $site) {
+            $isSiteReadOnly = $site.ReadOnly
+            $web = Get-SPWeb -Identity $WebID -Site $site -Limit ALL -ErrorAction SilentlyContinue
+            if ($null -ne $web) {
+                $webAppUrl = ($site.WebApplication.Url).TrimEnd('/')
+                $pageUrl = "{0}/{1}/{2}" -f $webAppUrl, $DirName, $LeafName
+                Write-Output "Page URL: $pageUrl"
+
+                if ($isSiteReadOnly) {
+                    Write-Output 'ReadOnly flag temporarily removed from site'
+                    $site.ReadOnly = $false
+                }
+
+                try {
+                    $spWebPartManager = $web.GetLimitedWebPartManager($pageUrl, [System.Web.UI.WebControls.WebParts.PersonalizationScope]::Shared)
+                    $webPartToDelete = $spWebPartManager.WebParts | Where-Object -FilterScript { $_.StorageKey -eq $StorageKey }
+                    if ($null -ne $webPartToDelete) {
+                        Write-Output "Removing WebPart with StorageKey $StorageKey from page"
+                        if ($PSCmdlet.ShouldProcess($pageUrl, "Delete WebPart with StorageKey $StorageKey")) {
+                            $spWebPartManager.DeleteWebPart($spWebPartManager.WebParts[$webPartToDelete.Id])
+                        }
+                    }
+                    else {
+                        Write-Output "WebPart with StorageKey $StorageKey not found on page"
+                    }
+                }
+                finally {
+                    if ($isSiteReadOnly) {
+                        Write-Output 'ReadOnly flag re-applied to site'
+                        $site.ReadOnly = $true
+                    }
+                    $web.Dispose()
+                }
+            }
+            else {
+                Write-Output "WebID $WebID does not exist.`nPlease check this WebID"
+            }
+            $site.Dispose()
+        }
+        else {
+            Write-Output "SiteID $SiteID does not exist.`nPlease check this SiteID"
+        }
+    }
+    catch {
+        return $_
+    }
+}
+
 function Remove-SPSMissingAssembly {
+    [CmdletBinding(SupportsShouldProcess = $true)]
     param
     (
         [Parameter()]
@@ -609,7 +812,9 @@ function Remove-SPSMissingAssembly {
                     }
                     if ($null -ne $AssemblyToDelete) {
                         Write-Output "Removing AssemblyID $AssemblyID from SPSite object."
-                        $AssemblyToDelete.delete()
+                        if ($PSCmdlet.ShouldProcess("SPSite $SiteID", "Delete AssemblyID $AssemblyID")) {
+                            $AssemblyToDelete.delete()
+                        }
                     }
                     else {
                         Write-Output "AssemblyID $AssemblyID does not exist in SPSite object."
@@ -629,7 +834,9 @@ function Remove-SPSMissingAssembly {
                     }
                     if ($null -ne $AssemblyToDelete) {
                         Write-Output "Removing AssemblyID $AssemblyID from SPWeb object."
-                        $AssemblyToDelete.delete()
+                        if ($PSCmdlet.ShouldProcess("SPWeb $WebID", "Delete AssemblyID $AssemblyID")) {
+                            $AssemblyToDelete.delete()
+                        }
                     }
                     else {
                         Write-Output "AssemblyID $AssemblyID does not exist in SPWeb object."
@@ -651,7 +858,9 @@ function Remove-SPSMissingAssembly {
                         }
                         if ($null -ne $AssemblyToDelete) {
                             Write-Output "Removing AssemblyID $AssemblyID from SPList object."
-                            $AssemblyToDelete.delete()
+                            if ($PSCmdlet.ShouldProcess("SPList $hostID", "Delete AssemblyID $AssemblyID")) {
+                                $AssemblyToDelete.delete()
+                            }
                         }
                         else {
                             Write-Output "AssemblyID $AssemblyID does not exist in SPList object."
@@ -674,6 +883,7 @@ function Remove-SPSMissingAssembly {
 }
 
 function Remove-SPSMissingConfiguration {
+    [CmdletBinding(SupportsShouldProcess = $true)]
     param
     (
         [Parameter()]
@@ -708,7 +918,9 @@ function Remove-SPSMissingConfiguration {
                         Write-Output "Removing login: $Login"
                         Write-Output "of SiteAdministrator Property of SPWeb:"
                         Write-Output "$($web.Url)"
-                        $web.SiteAdministrators.Remove($Login)
+                        if ($PSCmdlet.ShouldProcess($web.Url, "Remove SiteAdministrator $Login")) {
+                            $web.SiteAdministrators.Remove($Login)
+                        }
                     }
                     else {
                         Write-Output "$Login does not exist in SiteAdministrators Property"
@@ -731,6 +943,7 @@ function Remove-SPSMissingConfiguration {
 }
 
 function Remove-SPSOrphanedSite {
+    [CmdletBinding(SupportsShouldProcess = $true)]
     param
     (
         [Parameter()]
@@ -752,7 +965,9 @@ function Remove-SPSOrphanedSite {
         $spContentDb = Get-SPContentDatabase $Database
         if ($null -ne $spContentDb) {
             Write-Output "Removing SPSite object $SiteID with the method ForceDeleteSite"
-            $spContentDb.ForceDeleteSite($siteID, $false, $false)
+            if ($PSCmdlet.ShouldProcess("SiteID $SiteID in database $Database", 'ForceDeleteSite')) {
+                $spContentDb.ForceDeleteSite($siteID, $false, $false)
+            }
         }
         else {
             Write-Output "SPContentDatabse $Database does not exist.`nPlease check this Database"
